Oracle을 사용할수 있도록 작업 ( 상수값 관련 오류 처리 및 mysql 소스 주석처리된 상태로 커밋 )

+++ SqlParserLib/Source/Oracle/PlSqlLexerBase.h
@@ -12,7 +12,7 @@ namespace antlrcpp_oracle {
 	public:
 		bool IsNewlineAtPos(int pos)
 		{
-			int la = _input->LA(pos);
+			size_t la = _input->LA(pos);
 			return la == -1 || la == '\n';
 		};
 	};
